add validation

+++ app/Http/Controllers/ProductController.php
@@ -4,29 +4,27 @@ namespace App\Http\Controllers;
 
 use App\Http\Requests\Product\StoreRequest;
 use App\Http\Requests\Product\UpdateRequest;
+use App\Http\Resources\Product\IndexResource;
 use App\Http\Resources\ProductCollection;
 use App\Http\Resources\ProductResource;
 use App\Models\Product;
 use Illuminate\Http\Request;
-use Illuminate\Routing\Controllers\HasMiddleware;
-use Override;
 
-class ProductController extends Controller implements HasMiddleware
+class ProductController extends Controller
 {
-    #[Override]
-    public static function middleware()
-    {
-        return [
-            // 'auth:sanctum'  => [
-            //     'except'    => ['index', 'show']
-            // ]
-        ];
-    }
 
     public function index(Request $request)
     {
+
+        if ($request->category_id && !is_numeric($request->category_id))
+            return response()->json(['message' => 'Параметр некорректен'], 400);
+
         $products = Product::with('category')
-            ->paginate($request->per_page ?? 15);
+            ->when($request->category_id, function ($query, $categoryId) {
+                return $query->where('category_id', $categoryId);
+            })
+            ->paginate($request->per_page ?? 15)
+            ->withQueryString();
 
         return new ProductCollection($products);
     }
@@ -40,23 +38,54 @@ class ProductController extends Controller implements HasMiddleware
             ->setStatusCode(201);
     }
 
-    public function show(Product $product)
+    public function show(string $id)
     {
-        return new ProductResource($product->load('category'));
+        try {
+            if (!is_numeric($id))
+                return response()->json(['message' => 'Параметр некорректен'], 400);
+
+            $product = Product::find($id);
+            if (!$product)
+                return response()->json(['message' => 'Запись не найдена'], 404);
+
+            return new IndexResource($product->load('category'));
+        } catch (\Throwable $e) {
+            return response()->json([
+                'message'   => 'Server error'
+            ], 500);
+        }
     }
 
-    public function update(UpdateRequest $request, Product $product)
+    public function update(UpdateRequest $request, string $id)
     {
+        if (!is_numeric($id))
+            return response()->json(['message'  => 'Параметр некорректен'], 400);
+
+        $product = Product::find($id);
+        if (!$product)
+            return response()->json(['message' => 'Запись не найдена'], 404);
+
         $product->update($request->validated());
-
         return new ProductResource($product->load('category'));
     }
 
-    public function destroy(Product $product)
+    public function destroy(string $id)
     {
-        $product->delete();
-        return response()->json([
-            'message'   => 'Товар удален'
-        ]);
+        try {
+            if (!is_numeric($id))
+                return response()->json(['message' => 'Параметр некорректен'], 400);
+            $product = Product::find($id);
+            if (!$product)
+                return response()->json(['message' => 'Запись не найдена'], 404);
+
+            $product->delete();
+            return response()->json([
+                'message'   => 'Товар удален'
+            ]);
+        } catch (\Throwable $e) {
+            return response()->json([
+                'message'   => 'Server error'
+            ], $e->getCode());
+        }
     }
 }
